+++ tensor/t_mult.cc
@@ -11,7 +11,7 @@ namespace ctce {
         const vector<IndexName>& sum_ids,
         IterGroup <triangular>& sum_itr,
         IterGroup<CopyIter>& cp_itr,
-		const vector<Integer>& tid,
+		const vector<size_t>& tid,
 		Multiplication& m) {
 
       //vector<Integer>& vtab = Table::value();
@@ -19,14 +19,14 @@ namespace ctce {
       const vector<IndexName>& a_ids = id2name(m.a_ids);//tA.name();
       const vector<IndexName>& b_ids = id2name(m.b_ids);//tB.name();
 
-      vector<Integer> vtab(IndexNum);
+      vector<size_t> vtab(IndexNum);
       for (int i = 0; i < tC.dim(); ++i) {
         vtab[c_ids[i]] = tid[i]; // values used in genAntiIter
       }
 
       IterGroup<antisymm> ext_itr;
       genAntiIter(vtab, ext_itr,tC,tA,tB);
-      vector<Integer> ext_vec, sum_vec;
+      vector<size_t> ext_vec, sum_vec;
       ext_itr.reset();
 
       while (ext_itr.next(ext_vec)) {
@@ -38,12 +38,12 @@ namespace ctce {
             vtab[c_ids[i]] = ext_vec[i];
             //tC.setValueByName(c_ids[i], vtab[c_ids[i]]);
           }
-          //const vector<Integer>& c_ids_v = tC.value();
-	  vector<Integer> c_ids_v(tC.dim());
+          //const vector<size_t>& c_ids_v = tC.value();
+	  vector<size_t> c_ids_v(tC.dim());
 	  for(int i=0; i<tC.dim(); i++) {
 	    c_ids_v[i] = vtab[c_ids[i]];
 	  }
-          Integer dimc = compute_size(c_ids_v);
+          size_t dimc = compute_size(c_ids_v);
 	  assert(dimc == compute_size(tid));
           if (dimc == 0) continue;
           double* buf_c_sort = new double[dimc];
@@ -57,7 +57,7 @@ namespace ctce {
 	    for(int i=0; i<sum_vec.size(); i++) {
 	      vtab[sum_ids[i]] = sum_vec[i];
 	    }
-	    vector<Integer> a_ids_v(tA.dim()), b_ids_v(tB.dim());
+	    vector<size_t> a_ids_v(tA.dim()), b_ids_v(tB.dim());
 	    for(int i=0; i<tA.dim(); i++) {
 	      assert(a_ids[i] < IndexNum);
 	      a_ids_v[i] = vtab[a_ids[i]];
@@ -70,20 +70,20 @@ namespace ctce {
             //   tA.setValueByName(sum_ids[i],sum_vec[i]);
             //   tB.setValueByName(sum_ids[i],sum_vec[i]);
             // }
-            // vector<Integer> a_ids_v = tA.value();
-            // vector<Integer> b_ids_v = tB.value();
+            // vector<size_t> a_ids_v = tA.value();
+            // vector<size_t> b_ids_v = tB.value();
             if(!tA.is_spatial_nonzero(a_ids_v))  continue;
             if(!tA.is_spin_nonzero(a_ids_v)) continue;
 
-	    vector<Integer> a_value_r, b_value_r;
+	    vector<size_t> a_value_r, b_value_r;
 	    tA.gen_restricted(a_ids_v, a_value_r);
 	    tB.gen_restricted(b_ids_v, b_value_r);
 
-            Integer dim_common = compute_size(sum_vec);
-            Integer dima = compute_size(a_ids_v); if (dima<=0) continue;
-            Integer dimb = compute_size(b_ids_v); if (dimb<=0) continue;
-            Integer dima_sort = dima / dim_common;
-            Integer dimb_sort = dimb / dim_common;
+            size_t dim_common = compute_size(sum_vec);
+            size_t dima = compute_size(a_ids_v); if (dima<=0) continue;
+            size_t dimb = compute_size(b_ids_v); if (dimb<=0) continue;
+            size_t dima_sort = dima / dim_common;
+            size_t dimb_sort = dimb / dim_common;
 
 	    setValue(m.a_ids, a_ids_v);
 	    setValue(m.b_ids, b_ids_v);
@@ -92,8 +92,8 @@ namespace ctce {
 
             double* buf_a = new double[dima];
             double* buf_a_sort = new double[dima];
-	    vector<Integer> a_svalue_r, b_svalue_r;
-	    vector<Integer> a_svalue, b_svalue;
+	    vector<size_t> a_svalue_r, b_svalue_r;
+	    vector<size_t> a_svalue, b_svalue;
 	    vector<IndexName> a_name;
 	    vector<IndexName> b_name;
 	    int a_sign = sortByValueThenExtSymGroup(m.a_ids, a_name, a_svalue, a_svalue_r);
@@ -103,7 +103,7 @@ namespace ctce {
 	    // tA.get(*d_a,a_svalue_r,buf_a,dima,*k_a_offset);
 	    tA.get(a_svalue_r,buf_a,dima);
             //tA.get(*d_a,buf_a,dima,*k_a_offset);
-	    vector<Integer> a_sort_ids = sort_ids(a_name, m.a_mem_pos);
+	    vector<size_t> a_sort_ids = sort_ids(a_name, m.a_mem_pos);
             tce_sort(buf_a, buf_a_sort, a_svalue /*tA._value()*/, a_sort_ids/*tA.sort_ids()*/, (double)a_sign /*tA.sign()*/);
             delete [] buf_a;
 
@@ -117,7 +117,7 @@ namespace ctce {
 	    tB.get(b_svalue_r,buf_b,dimb);
             //tB.get(*d_b,buf_b,dimb,*k_b_offset);
             //tce_sort(buf_b, buf_b_sort, tB._value(), tB.sort_ids(), (double)tB.sign());
-	    vector<Integer> b_sort_ids = sort_ids(b_name, m.b_mem_pos);
+	    vector<size_t> b_sort_ids = sort_ids(b_name, m.b_mem_pos);
 	    tce_sort(buf_b, buf_b_sort, b_svalue/*tB._value()*/, b_sort_ids /*tB.sort_ids(b_name)*/, (double)b_sign /*(double)tB.sign()*/);
             delete [] buf_b;
 
@@ -132,20 +132,20 @@ namespace ctce {
 	  setValueR(m.c_ids, c_ids_v);
           //tC.sortByValueThenExtSymGroup();
 	  vector<IndexName> c_name;
-	  vector<Integer> c_svalue_r, c_svalue;
+	  vector<size_t> c_svalue_r, c_svalue;
 	  sortByValueThenExtSymGroup(m.c_ids, c_name, c_svalue, c_svalue_r);
           cp_itr.reset();
-          vector<Integer> perm;
+          vector<size_t> perm;
           while (cp_itr.next(perm)) {
             cp_itr.fix_ids_for_copy(perm);
 	    vector<IndexName> name;
-	    vector<Integer> value, value_r;
+	    vector<size_t> value, value_r;
             //tC.orderIds(perm);
 	    orderIds(m.c_ids, perm, name, value, value_r);
-            if (compareVec<Integer>(tid,value)) {
+            if (compareVec<size_t>(tid,value)) {
               double sign = coef * cp_itr.sign();
-	      std::vector<Integer> cperm = mult_perm(name, m.c_mem_pos);
-	      std::vector<Integer> cmpval = getMemPosVal(m.c_ids, m.c_mem_pos);
+	      std::vector<size_t> cperm = mult_perm(name, m.c_mem_pos);
+	      std::vector<size_t> cmpval = getMemPosVal(m.c_ids, m.c_mem_pos);
               sortacc(buf_c_sort, a_c, cmpval, cperm, sign);
             }
           } // cp_itr
@@ -164,14 +164,14 @@ namespace ctce {
 		 Multiplication& m) {
 
       const vector<IndexName>& c_name = id2name(m.c_ids);//tC.name();
-      vector<Integer> out_vec;
+      vector<size_t> out_vec;
       out_itr.reset();
       double anti=0, add=0;
       while (out_itr.next(out_vec)) {
         if (!tC.is_spatial_nonzero(out_vec)) continue;
         if (!tC.is_spin_nonzero(out_vec)) continue;
         if (!tC.is_spin_restricted_nonzero(out_vec)) continue;
-        Integer dimc = compute_size(out_vec);
+        size_t dimc = compute_size(out_vec);
         if (dimc <= 0) continue;
         double* buf_c = new double[dimc];
         memset(buf_c, 0, dimc * sizeof(double));
@@ -192,7 +192,7 @@ namespace ctce {
 		 IterGroup<triangular>& out_itr,
 		 Multiplication& m, int sync_ga, int spos) {
 
-      //vector<Integer>& vtab = Table::value();
+      //vector<size_t>& vtab = Table::value();
       const vector<IndexName>& c_ids = id2name(m.c_ids);//tC.name();
       const vector<IndexName>& a_ids = id2name(m.a_ids);//tA.name();
       const vector<IndexName>& b_ids = id2name(m.b_ids);//tB.name();
@@ -217,7 +217,7 @@ namespace ctce {
 
       int next = NGA_Read_inc(taskHandle, &sub, 1);
 
-      vector<Integer> out_vec, sum_vec;
+      vector<size_t> out_vec, sum_vec;
       out_itr.reset();
 
       bool out_once = true;
@@ -238,9 +238,9 @@ namespace ctce {
 	      assert(c_ids[i] < IndexNum);
 	      vtab1[c_ids[i]] = out_vec[i];
             }
-	    vector<Integer> c_ids_v = out_vec;
+	    vector<size_t> c_ids_v = out_vec;
 
-            Integer dimc = compute_size(c_ids_v); if (dimc<=0) continue;
+            size_t dimc = compute_size(c_ids_v); if (dimc<=0) continue;
             double* buf_c_sort = new double[dimc];
             memset(buf_c_sort, 0, dimc*sizeof(double));
 
@@ -253,7 +253,7 @@ namespace ctce {
 	      for(int i=0; i<sum_vec.size(); i++) {
 		vtab1[sum_ids[i]] = sum_vec[i];
 	      }
-	      vector<Integer> a_ids_v(tA.dim()), b_ids_v(tB.dim());
+	      vector<size_t> a_ids_v(tA.dim()), b_ids_v(tB.dim());
 	      for(int i=0; i<tA.dim(); i++) {
 		assert(a_ids[i] < IndexNum);
 		a_ids_v[i] = vtab1[a_ids[i]];
@@ -272,15 +272,15 @@ namespace ctce {
 	      }
 #endif
 	      
-	      vector<Integer> a_value_r, b_value_r;
+	      vector<size_t> a_value_r, b_value_r;
 	      tA.gen_restricted(a_ids_v, a_value_r);
 	      tB.gen_restricted(b_ids_v, b_value_r);
 
-              Integer dim_common = compute_size(sum_vec);
-              Integer dima = compute_size(a_ids_v); if (dima<=0) continue;
-              Integer dimb = compute_size(b_ids_v); if (dimb<=0) continue;
-              Integer dima_sort = dima / dim_common;
-              Integer dimb_sort = dimb / dim_common;
+              size_t dim_common = compute_size(sum_vec);
+              size_t dima = compute_size(a_ids_v); if (dima<=0) continue;
+              size_t dimb = compute_size(b_ids_v); if (dimb<=0) continue;
+              size_t dima_sort = dima / dim_common;
+              size_t dimb_sort = dimb / dim_common;
 
               double* buf_a = new double[dima];
               double* buf_a_sort = new double[dima];
@@ -294,8 +294,8 @@ namespace ctce {
 	      setValueR(m.a_ids, a_value_r);
 	      setValueR(m.b_ids, b_value_r);
 
-	      vector<Integer> a_svalue_r, b_svalue_r;
-	      vector<Integer> a_svalue, b_svalue;
+	      vector<size_t> a_svalue_r, b_svalue_r;
+	      vector<size_t> a_svalue, b_svalue;
 	      vector<IndexName> a_name;
 	      vector<IndexName> b_name;
               // int a_sign = tA.sortByValueThenExtSymGroup(a_name, a_svalue, a_svalue_r);
@@ -322,8 +322,8 @@ namespace ctce {
               // tA.get(*d_a,a_svalue_r,buf_a,dima,*k_a_offset);
               tA.get(a_svalue_r,buf_a,dima);
               getTimer.stop();
-	      vector<Integer> a_sort_ids = sort_ids(a_name, m.a_mem_pos);
-	      //vector<Integer> a_sort_ids = tA.sort_ids(a_name);
+	      vector<size_t> a_sort_ids = sort_ids(a_name, m.a_mem_pos);
+	      //vector<size_t> a_sort_ids = tA.sort_ids(a_name);
 
               tce_sort(buf_a, buf_a_sort, a_svalue/*tA._value()*/, a_sort_ids/*tA.sort_ids(a_name)*/, (double)a_sign /*(double)tA.sign()*/);
               delete [] buf_a;
@@ -336,8 +336,8 @@ namespace ctce {
               // tB.get(*d_b,b_svalue_r,buf_b,dimb,*k_b_offset);
               tB.get(b_svalue_r,buf_b,dimb);
               getTimer.stop();
-	      //vector<Integer> b_sort_ids = tB.sort_ids(b_name);
-	      vector<Integer> b_sort_ids = sort_ids(b_name, m.b_mem_pos);
+	      //vector<size_t> b_sort_ids = tB.sort_ids(b_name);
+	      vector<size_t> b_sort_ids = sort_ids(b_name, m.b_mem_pos);
               tce_sort(buf_b, buf_b_sort, b_svalue/*tB._value()*/, b_sort_ids /*tB.sort_ids(b_name)*/, (double)b_sign /*(double)tB.sign()*/);
               delete [] buf_b;
 
@@ -359,29 +359,29 @@ namespace ctce {
 	    setValueR(m.c_ids, c_ids_v);
 #endif
 	  vector<IndexName> c_name;
-	  vector<Integer> c_svalue_r, c_svalue;
+	  vector<size_t> c_svalue_r, c_svalue;
 	  // tC.sortByValueThenExtSymGroup(c_name, c_svalue, c_svalue_r);
 	  sortByValueThenExtSymGroup(m.c_ids, c_name, c_svalue, c_svalue_r);
-	  //vector<Integer> tid = tC._value();
-	  vector<Integer> tid = c_svalue;
+	  //vector<size_t> tid = tC._value();
+	  vector<size_t> tid = c_svalue;
 
             cp_itr.reset();
-            vector<Integer> perm;
+            vector<size_t> perm;
             bool out_cp_one_time = true;
             while (cp_itr.next(perm) || out_cp_one_time) {
               out_cp_one_time = false;
               cp_itr.fix_ids_for_copy(perm);
 	      vector<IndexName> name;
-	      vector<Integer> value, value_r;
+	      vector<size_t> value, value_r;
               //tC.orderIds(perm, name, value, value_r);
 	      orderIds(m.c_ids, perm, name, value, value_r);
-              if (compareVec<Integer>(tid, value)) {
+              if (compareVec<size_t>(tid, value)) {
                 double sign = coef * cp_itr.sign();
                 double* buf_c = new double[dimc];
-		//std::vector<Integer> cperm = tC.perm(name);
-		std::vector<Integer> cperm = mult_perm(name, m.c_mem_pos);
-		// std::vector<Integer> cmpval = getMemPosVal(tC.ids(), m.c_mem_pos);
-		std::vector<Integer> cmpval = getMemPosVal(m.c_ids, m.c_mem_pos);
+		//std::vector<size_t> cperm = tC.perm(name);
+		std::vector<size_t> cperm = mult_perm(name, m.c_mem_pos);
+		// std::vector<size_t> cmpval = getMemPosVal(tC.ids(), m.c_mem_pos);
+		std::vector<size_t> cmpval = getMemPosVal(m.c_ids, m.c_mem_pos);
                 tce_sort(buf_c_sort, buf_c, cmpval, cperm, sign);
 
                 addTimer.start();
